feat(svelte-system-ui-components): add Flex component
