+++ libs/svelte-system-ui-components/src/components/Flex.svelte
@@ -1,0 +1,16 @@
+<script>
+  import { styled } from 'svelte-system-ui'
+  import theme from '../theme'
+  import Box from './Box'
+
+  export let style
+  let compStyles = {
+    d: 'flex',
+    ...style,
+  }
+  console.log('Flex', compStyles, $$props)
+</script>
+
+<Box {...$$props} style={compStyles}>
+  <slot />
+</Box>
