+++ libs/svelte-system-ui-components/src/components/stories/box.stories.js
@@ -2,7 +2,7 @@ import { action } from '@storybook/addon-actions'
 
 import Box from '../Box.svelte'
 import BoxView from './views/boxView.svelte'
-import BoxViewFlex from './views/boxViewFlex.svelte'
+import BoxViewChildren from './views/boxViewChildren.svelte'
 
 import { basic } from '../../theme'
 
@@ -27,20 +27,8 @@ export const withBackground = () => ({
   props: {
     style: {
       bg: 'primary',
-      p: [5, 3, 1],
-      m: [5, 0, 1],
-      color: 'muted',
-      brd: `3px solid`,
-      brdcol: `secondary`,
-      _hover: { bg: 'secondary' },
-      _after: {
-        content: '"→"',
-        pl: '20px',
-      },
-      fontFamily: 'heading',
     },
     boxText: 'Box with theme primary color background',
-    role: 'header',
   },
 })
 
@@ -69,6 +57,8 @@ export const withPaddingLarge = () => ({
   props: {
     style: {
       p: 4,
+      bg: 'primary',
+      color: 'background',
     },
     boxText: 'Box with large padding',
   },
@@ -87,7 +77,7 @@ export const withBorder = () => ({
 })
 
 export const withDisplayBlock = () => ({
-  Component: BoxViewFlex,
+  Component: BoxViewChildren,
   props: {
     style: { p: 4, brd: '1px solid', brdCol: 'primary' },
     style1: { p: 4, brd: '1px solid', brdCol: 'primary' },
@@ -100,7 +90,7 @@ export const withDisplayBlock = () => ({
 })
 
 export const withDisplayInline = () => ({
-  Component: BoxViewFlex,
+  Component: BoxViewChildren,
   props: {
     style: {
       p: 4,
@@ -128,273 +118,5 @@ export const withDisplayInline = () => ({
     boxText1: 'Just nested',
     boxText2: 'but not using flex;',
     boxText3: 'set to display as "inline"',
-  },
-})
-
-export const withFlex = () => ({
-  Component: BoxViewFlex,
-  props: {
-    style: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'primary',
-      d: 'flex',
-    },
-    style1: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    style2: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    style3: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    boxText1: 'Nested',
-    boxText2: 'in parent',
-    boxText3: 'using flex',
-  },
-})
-
-export const withFlexColumn = () => ({
-  Component: BoxViewFlex,
-  props: {
-    style: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'primary',
-      d: 'flex',
-      flexdir: 'column',
-    },
-    style1: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    style2: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    style3: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    boxText1: '1: Nested',
-    boxText2: '2: in flex parent',
-    boxText3: '3: flex-direction: column',
-  },
-})
-
-export const withFlexRowReverse = () => ({
-  Component: BoxViewFlex,
-  props: {
-    style: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'primary',
-      d: 'flex',
-      flexdir: 'row-reverse',
-    },
-    style1: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    style2: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    style3: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    boxText1: '1: Nested',
-    boxText2: '2: in flex parent',
-    boxText3: '3: flex-direction: row-reverse',
-  },
-})
-
-export const withFlexOrder = () => ({
-  Component: BoxViewFlex,
-  props: {
-    style: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'primary',
-      d: 'flex',
-      flexdir: 'column',
-    },
-    style1: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'secondary',
-      ord: '1',
-    },
-    style2: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'secondary',
-      ord: '2',
-    },
-    style3: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'secondary',
-      ord: '0',
-    },
-    boxText1: '1: Nested (order: 1)',
-    boxText2: '2: in flex parent (order: 2)',
-    boxText3: '3: using order (order: 0) ',
-  },
-})
-
-export const withFlexGrow = () => ({
-  Component: BoxViewFlex,
-  props: {
-    style: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'primary',
-      d: 'flex',
-      flexdir: 'row',
-    },
-    style1: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    style2: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'secondary',
-      flexGrow: '2',
-    },
-    style3: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    boxText1: '1: normal',
-    boxText2: '2: I am sooo big',
-    boxText3: '3: normal',
-  },
-})
-
-export const withFlexShrink = () => ({
-  Component: BoxViewFlex,
-  props: {
-    style: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'primary',
-      d: 'flex',
-      flexdir: 'row',
-    },
-    style1: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    style2: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'secondary',
-      flexshr: '2',
-    },
-    style3: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    boxText1:
-      '1: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
-    boxText2: '2: I AM SHRUNK-ED. Lorem ipsum dolor sit amet...',
-    boxText3:
-      '3: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
-  },
-})
-
-export const withFlexBasis = () => ({
-  Component: BoxViewFlex,
-  props: {
-    style: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'primary',
-      d: 'flex',
-      flexdir: 'row',
-    },
-    style1: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    style2: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'secondary',
-      flexbas: '50%',
-    },
-    style3: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    boxText1:
-      '1: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
-    boxText2: '2: I AM THE BASIS. Lorem ipsum dolor sit amet...',
-    boxText3:
-      '3: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
-  },
-})
-
-export const withAlignItemsAndSelf = () => ({
-  Component: BoxViewFlex,
-  props: {
-    style: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'primary',
-      d: 'flex',
-      flexdir: 'column',
-      align: 'center',
-    },
-    style1: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'secondary',
-      slf: 'flex-end',
-    },
-    style2: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    style3: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    boxText1: '1: I want to be different',
-    boxText2: '2: just a box',
-    boxText3: '3: just a box',
-  },
-})
-
-export const withFlexWrap = () => ({
-  Component: BoxViewFlex,
-  props: {
-    style: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'primary',
-      d: 'flex',
-      flexWrap: 'wrap',
-    },
-    style1: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'secondary',
-    },
-    style2: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    style3: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    boxText1: '1: Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
-    boxText2: '2: Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
-    boxText3:
-      '3: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
-  },
-})
-
-export const withJustifySpaceBetween = () => ({
-  Component: BoxViewFlex,
-  props: {
-    style: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'primary',
-      d: 'flex',
-      flexdir: 'row',
-      justC: 'space-between',
-    },
-    style1: {
-      p: 4,
-      brd: '1px solid',
-      brdCol: 'secondary',
-      slf: 'flex-end',
-    },
-    style2: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    style3: { p: 4, brd: '1px solid', brdCol: 'secondary' },
-    boxText1: '1: Stay',
-    boxText2: '2: away',
-    boxText3: '3: from me',
-  },
-})
-
-export const withResponsive = () => ({
-  Component: BoxViewFlex,
-  props: {
-    style: {
-      bg: '#ccc',
-      color: 'background',
-      d: 'flex',
-      flexWrap: 'wrap',
-      flexdir: 'row',
-      fontSize: ['20px', '15px', '40px'],
-    },
-    style1: {
-      bg: 'red',
-      w: [1, 1 / 2, 1 / 3],
-      h: '20rem',
-    },
-    style2: {
-      bg: 'blue',
-      w: [1, 1 / 2, 1 / 3],
-      h: '20rem',
-    },
-    style3: { bg: 'green', w: [1, 1, 1 / 3], h: '20rem' },
-    boxText1: 'Font size',
-    boxText2: 'is reponsive',
-    boxText3: 'too!',
   },
 })
